Started todolist
Hello GitHub!

I started work on todolist tonight, a very simple app with an MIT license to help other beginner coders like me learn the basics of code!

Started to make this an Expo React Native project but Expo Go was giving me issues so I just coded very basic React Native code and called it a night.

Learned CMD A + CMD + / on VSCODE so that was cool!

Goodnight GitHub!

.j.